finish becomeProvider

+++ CRP/CRP/Areas/Customer/Controllers/BookingController.cs
@@ -122,13 +122,13 @@ namespace CRP.Areas.Customer.Controllers
 		}
 
 		// Handler for TryBookingApi
-		private void CheckPendingBooking(int bookingID)
+		private static void CheckPendingBooking(int bookingID)
 		{
 			var dbContext = new CRPEntities();
 			var bookingService = new BookingReceiptService(new UnitOfWork(dbContext), new BookingReceiptRepository(dbContext));
 			var bookingReceipt = bookingService.Get(bookingID);
 
-			if(bookingReceipt.IsPending)
+			if(bookingReceipt != null && bookingReceipt.IsPending)
 				bookingService.Delete(bookingReceipt);
 		}
 
@@ -145,7 +145,7 @@ namespace CRP.Areas.Customer.Controllers
 			if (bookingReceipt == null)
 				return new HttpStatusCodeResult(403, "Access denied.");
 
-			var bookingConfirmViewModel = new BookingConfirmViewModel {Receipt = bookingReceipt, NganLuong = new NganLuongBookingModel()};
+			var bookingConfirmViewModel = new BookingConfirmViewModel {Receipt = bookingReceipt, NganLuong = new NganLuongPaymentModel()};
 			bookingConfirmViewModel.NganLuong.OrderCode = bookingReceipt.ID.ToString();
 
 			return View("~/Areas/Customer/Views/Booking/BookingConfirm.cshtml", bookingConfirmViewModel);
@@ -156,20 +156,24 @@ namespace CRP.Areas.Customer.Controllers
 		[System.Web.Http.HttpPost]
 		[ValidateAntiForgeryToken]
 		[System.Web.Mvc.Route("bookingConfirm", Name = "BookVehicle")]
-		public System.Web.Mvc.ActionResult BookVehicle(BookingConfirmViewModel bookingModel)
+		public System.Web.Mvc.ActionResult BookVehicle(BookingConfirmViewModel bookingModel, NganLuongPaymentModel nganLuongPayment)
 		{
+			// Check if the request contains all valid params
+			if (bookingModel?.Action == null || bookingModel.Receipt?.ID == null || nganLuongPayment == null)
+				return new HttpStatusCodeResult(400, "Invalid request");
+
 			var bookingService = this.Service<IBookingReceiptService>();
-			var BookingReceipt = bookingService.Get(bookingModel.Receipt.ID);
+			var bookingReceipt = bookingService.Get(bookingModel.Receipt.ID);
 
 			// Act based on the received action's name
 			switch (bookingModel.Action)
 			{
 				case "delete":
-					bookingService.Delete(BookingReceipt);
+					bookingService.Delete(bookingReceipt);
 					return RedirectToAction("Index", "Home");
 				case "change":
-					var vehicleID = BookingReceipt.VehicleID;
-					bookingService.Delete(BookingReceipt);
+					var vehicleID = bookingReceipt.VehicleID;
+					bookingService.Delete(bookingReceipt);
 					return RedirectToAction("VehicleInfo", "Home", new {id = vehicleID});
 				case "pay":
 					break;
@@ -180,25 +184,24 @@ namespace CRP.Areas.Customer.Controllers
 			// Only "pay" action left to handle
 			// Now validate nganluong params before redirect to nganluong
 
-			var info = new RequestInfoTestTemplate()
-			{
-				bank_code = bookingModel.NganLuong.BankCode,
-				Order_code = bookingModel.NganLuong.OrderCode,
-				order_description = "Test booking",
-				return_url = "http://localhost:65358/bookingReceipt",
-				cancel_url = "http://localhost:65358/bookingReceipt?canceledBookingID=" + bookingModel.Receipt.ID
-			};
-
 			var user = HttpContext.GetOwinContext()
 					.GetUserManager<ApplicationUserManager>()
 					.FindById(HttpContext.User.Identity.GetUserId());
 
-			info.Buyer_fullname = user.FullName;
-			info.Buyer_email = user.Email;
-			info.Buyer_mobile = user.PhoneNumber;
-
+			var info = new RequestInfoTestTemplate
+			{
+				bank_code = nganLuongPayment.BankCode,
+				Order_code = nganLuongPayment.OrderCode,
+				order_description = "Test booking",
+				return_url = "http://localhost:65358/bookingReceipt",
+				cancel_url = "http://localhost:65358/bookingReceipt?canceledBookingID=" + bookingModel.Receipt.ID,
+				Buyer_fullname = user.FullName,
+				Buyer_email = user.Email,
+				Buyer_mobile = user.PhoneNumber
+			};
+			
 			var objNLCheckout = new APICheckoutV3();
-			var result = objNLCheckout.GetUrlCheckout(info, bookingModel.NganLuong.PaymentMethod);
+			var result = objNLCheckout.GetUrlCheckout(info, nganLuongPayment.PaymentMethod);
 
 			if (result.Error_code == "00")
 			{
@@ -231,8 +234,7 @@ namespace CRP.Areas.Customer.Controllers
 			}
 			
 			// If the transaction went smoothy, check the returned info + MD5 token
-			var info = new RequestCheckOrderTestTemplate();
-			info.Token = token;
+			var info = new RequestCheckOrderTestTemplate {Token = token};
 			var objNLCheckout = new APICheckoutV3();
 			var result = objNLCheckout.GetTransactionDetail(info);
 
@@ -267,33 +269,6 @@ namespace CRP.Areas.Customer.Controllers
 			}
 
 			return new HttpStatusCodeResult(400, "Invalid request");
-
-			//SystemService sysService = new SystemService();
-			//lastBooking = 1;
-			//var service = this.Service<IBookingReceiptService>();
-			//var entity = await service.GetAsync(lastBooking);
-			////kiem tra xem da thanh toan thanh cong hay chua
-			//Boolean paySuccess = true;
-			////xuong databse ispending = false neu thanh toan thanh cong, xoa booking neu no ko thanh cong
-			////neu thanh toan thanh cong
-			//if (paySuccess)
-			//{
-			//	entity.IsPending = false;
-			//	service.Update(entity);
-			//	//tra ve model cua entity booking moi nhat
-
-			//}
-			//else
-			//{
-			//	service.Delete(entity);
-			//	//stop stread sau 5p kiem tra
-			//	DeleteBookingThread = true;
-			//	ViewBag.ErrorForPayment = "Thanh toan khong thanh cong!";
-			//}
-
-			//send mail bao cho Provider va Customer
-			//sysService.SendMailBooking("tamntse61384@fpt.edu.vn", entity);
-			//sysService.SendMailBooking(entity.AspNetUser11.Email, entity);
 		}
 
 
@@ -318,8 +293,8 @@ namespace CRP.Areas.Customer.Controllers
 			String customerID = User.Identity.GetUserId();
 			var service = this.Service<IBookingReceiptService>();
 			var list = service.GetBookingReceiptWithUser(customerID);
-			DateTime now = System.DateTime.Now;
-			foreach(BookingReceipt item in list.ToList())
+			var now = System.DateTime.Now;
+			foreach(var item in list.ToList())
 			{
 				if (item.EndTime < now)
 				{
